requirements.txt
ocrmypdf==16.3.1
openai==1.34.0
opencv-python-headless==4.10.0.84
pytesseract==0.3.10
tabula-py==2.9.3
tabulate==0.9.0
tesseract==0.1.3